move epics_names function to epics data file and update paths
