+++ frontend/pages/data/epics.py
@@ -1,0 +1,14 @@
+import requests
+from typing import List
+from config import base_url
+from pages.data.data import Select
+
+
+def epics_names() -> List[Select]:
+    api_epic_name = f"{base_url}/api/epics/active"
+    response_epic_name = requests.get(api_epic_name)
+    epic_name_rows = [Select(value="", display_value="select epic")]
+    for item in response_epic_name.json():
+        d = Select(value=item["id"], display_value=item["name"])
+        epic_name_rows.append(d)
+    return epic_name_rows
